requirements.txt
alabaster==0.7.9
appdirs==1.4.0
argcomplete==1.8.1
astroid==1.4.9
Babel==2.3.4
beautifulsoup4==4.5.3
bleach==1.5.0
chardet==2.3.0
decorator==4.0.11
docutils==0.13.1
docx2txt==0.6
EbookLib==0.15
entrypoints==0.2.2
html5lib==0.9999999
imagesize==0.7.1
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
isort==4.2.5
jedi==0.9.0
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
lazy-object-proxy==1.2.2
lxml==3.7.2
MarkupSafe==0.23
mccabe==0.6.0
mistune==0.7.3
mysql-connector==2.1.4
nbconvert==5.1.1
nbformat==4.2.0
notebook==4.3.1
numpydoc==0.6.0
olefile==0.44
packaging==16.8
pandocfilters==1.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
protobuf==3.1.0.post1
psutil==5.0.1
ptyprocess==0.5.1
pyflakes==1.5.0
Pygments==2.1.3
pylint==1.6.5
pyparsing==2.1.10
PyQt5==5.7.1
PySocks==1.6.5
pytesseract==0.1.6
python-pptx==0.6.2
pytz==2016.10
pyzmq==16.0.2
QtAwesome==0.4.3
qtconsole==4.2.1
QtPy==1.2.1
requests==2.13.0
rope-py3k==0.9.4.post1
simplegeneric==0.8.1
sip==4.19
six==1.10.0
snowballstemmer==1.2.1
SpeechRecognition==3.6.0
Sphinx==1.5.2
spyder==3.1.2
terminado==0.6
testpath==0.3
textract==1.5.0
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
widgetsnbextension==1.2.6
wrapt==1.10.8
xlrd==1.0.0
XlsxWriter==0.9.6
